split qwen3-next alloc-fail and deepep crossings to extra-b
