+++ test/registered/models_e2e/test_qwen3_next_models_extra.py
@@ -1,0 +1,83 @@
+"""Mamba lazy-alloc fault injection for Qwen3-Next (extra-b).
+
+Split out of `test_qwen3_next_models.py`: both classes force
+`SGLANG_TEST_MAMBA_LAZY_ALLOC_FAIL`, a test-only fault-injection path, while
+the mainline page_size=1/2 coverage stays per-commit in the parent file.
+"""
+
+import os
+import unittest
+
+from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.kits.kl_divergence_kit import KLDivergenceMixin
+from sglang.test.server_fixtures.default_fixture import DefaultServerBase
+
+register_cuda_ci(est_time=250, stage="extra-b", runner_config="4-gpu-h100")
+
+QWEN3_NEXT_MODEL = "Qwen/Qwen3-Next-80B-A3B-Instruct"
+
+_COMMON_ARGS = [
+    "--trust-remote-code",
+    "--tp-size",
+    "4",
+    "--chunked-prefill-size",
+    "2048",
+    "--mamba-scheduler-strategy",
+    "extra_buffer_lazy",
+    "--attention-backend",
+    "triton",
+]
+
+
+def _make_args(*, page_size=1, track_interval=2):
+    return [
+        *_COMMON_ARGS,
+        "--mamba-track-interval",
+        str(track_interval),
+        "--page-size",
+        str(page_size),
+    ]
+
+
+class TestQwen3NextLazyExtraBufferAllocFail(KLDivergenceMixin, DefaultServerBase):
+    model = QWEN3_NEXT_MODEL
+    cache_chunk_size = 64
+    kl_div_thres = 0.002
+    other_args = _make_args(page_size=1, track_interval=2)
+
+    @classmethod
+    def setUpClass(cls):
+        os.environ["SGLANG_TEST_MAMBA_LAZY_ALLOC_FAIL"] = "1"
+        os.environ["SGLANG_TEST_SKIP_CACHE_HIT_ASSERT"] = "1"
+        super().setUpClass()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass()
+        os.environ.pop("SGLANG_TEST_MAMBA_LAZY_ALLOC_FAIL", None)
+        os.environ.pop("SGLANG_TEST_SKIP_CACHE_HIT_ASSERT", None)
+
+
+class TestQwen3NextLazyExtraBufferLargePageAllocFail(
+    KLDivergenceMixin, DefaultServerBase
+):
+    model = QWEN3_NEXT_MODEL
+    cache_chunk_size = 64
+    kl_div_thres = 0.002
+    other_args = _make_args(page_size=2, track_interval=2)
+
+    @classmethod
+    def setUpClass(cls):
+        os.environ["SGLANG_TEST_MAMBA_LAZY_ALLOC_FAIL"] = "1"
+        os.environ["SGLANG_TEST_SKIP_CACHE_HIT_ASSERT"] = "1"
+        super().setUpClass()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass()
+        os.environ.pop("SGLANG_TEST_MAMBA_LAZY_ALLOC_FAIL", None)
+        os.environ.pop("SGLANG_TEST_SKIP_CACHE_HIT_ASSERT", None)
+
+
+if __name__ == "__main__":
+    unittest.main()
